fix: address review comments - ref title, Triandis name, route slug rename
Agent-Logs-Url: https://github.com/clarkemoyer/technologyadoptionbarriers.org/sessions/96ba7086-ba7d-490d-aea7-8a167fdd1bc3
